+++ R/helpers_ggscatterstats_subtitles.R
@@ -5,30 +5,30 @@
 #' @param type Type of association between paired samples required
 #'   ("`"parametric"`: Pearson's product moment correlation coefficient" or
 #'   "`"nonparametric"`: Spearman's rho" or "`"robust"`: percentage bend
-#'   correlation coefficient"). Corresponding abbreviations are also accepted:
-#'   `"p"` (for parametric/pearson's), `"np"` (nonparametric/spearman), `"r"`
-#'   (robust), resp.
+#'   correlation coefficient" or "`"bayes"`: Bayes Factor for Pearson's *r*").
+#'   Corresponding abbreviations are also accepted: `"p"` (for
+#'   parametric/pearson's), `"np"` (nonparametric/spearman), `"r"` (robust),
+#'   `"bf"` (for bayes factor), resp.
 #' @param messages Decides whether messages references, notes, and warnings are
 #'   to be displayed (Default: `TRUE`).
 #' @inheritParams robcor_ci
 #' @inheritParams cor_test_ci
 #' @inheritParams groupedstats::specify_decimal_p
+#' @inheritParams bf_corr_test
 #'
 #' @importFrom dplyr select
-#' @importFrom magrittr "%<>%"
-#' @importFrom magrittr "%>%"
 #' @importFrom rlang enquo
 #' @importFrom stats cor.test
 #'
 #' @examples
-#' 
+#'
 #' # without changing defaults
 #' subtitle_ggscatterstats(
 #'   data = ggplot2::midwest,
 #'   x = area,
 #'   y = percblack
 #' )
-#' 
+#'
 #' # changing defaults
 #' subtitle_ggscatterstats(
 #'   data = ggplot2::midwest,
@@ -48,6 +48,7 @@ subtitle_ggscatterstats <-
              nboot = 100,
              beta = 0.1,
              type = "pearson",
+             bf.prior = 0.707,
              conf.level = 0.95,
              conf.type = "norm",
              messages = TRUE,
@@ -61,20 +62,23 @@ subtitle_ggscatterstats <-
       ) %>%
       stats::na.omit(.)
 
-    #-------------------------------------------------- Pearson's r -----------------------------------------------------
-    #
-    if (type == "pearson" || type == "p") {
+    # the total sample size for analysis
+    sample_size <- nrow(x = data)
 
-      # Pearson's r
-      c <-
-        stats::cor.test(
-          formula = ~x + y,
-          data = data,
-          method = "pearson",
-          alternative = "two.sided",
-          exact = FALSE,
-          na.action = na.omit
-        )
+    # Pearson's r (will also be used for Bayes tests)
+    pearson_r_res <-
+      stats::cor.test(
+        formula = ~x + y,
+        data = data,
+        method = "pearson",
+        alternative = "two.sided",
+        exact = FALSE,
+        na.action = na.omit
+      )
+
+    #------------------------ Pearson's r --------------------------------------
+    #
+    if (type == "pearson" || type == "parametric" || type == "p") {
 
       # preparing the label
       subtitle <-
@@ -103,18 +107,21 @@ subtitle_ggscatterstats <-
             ),
           env = base::list(
             # degrees of freedom are always integer
-            df = c$parameter[[1]],
-            t = ggstatsplot::specify_decimal_p(x = c$statistic[[1]], k),
-            estimate = ggstatsplot::specify_decimal_p(x = c$estimate[[1]], k),
-            LL = ggstatsplot::specify_decimal_p(x = c$conf.int[1][[1]], k),
-            UL = ggstatsplot::specify_decimal_p(x = c$conf.int[2][[1]], k),
-            pvalue = ggstatsplot::specify_decimal_p(x = c$p.value[[1]], k, p.value = TRUE),
-            n = nrow(x = data)
+            df = pearson_r_res$parameter[[1]],
+            estimate = ggstatsplot::specify_decimal_p(x = pearson_r_res$estimate[[1]], k),
+            LL = ggstatsplot::specify_decimal_p(x = pearson_r_res$conf.int[1][[1]], k),
+            UL = ggstatsplot::specify_decimal_p(x = pearson_r_res$conf.int[2][[1]], k),
+            pvalue = ggstatsplot::specify_decimal_p(
+              x = pearson_r_res$p.value[[1]],
+              k = k,
+              p.value = TRUE
+            ),
+            n = sample_size
           )
         )
 
-      #-------------------------------------------------- Spearnman's rho -----------------------------------------------------
-    } else if (type == "spearman" || type == "np") {
+      #--------------------- Spearnman's rho ------------------------------------
+    } else if (type == "spearman" || type == "nonparametric" || type == "np") {
 
       # note that stats::cor.test doesn't give degress of freedom; it's
       # calculated as df = (no. of pairs - 2)
@@ -177,7 +184,7 @@ subtitle_ggscatterstats <-
               n
             ),
           env = base::list(
-            df = (length(data$x) - 2),
+            df = (sample_size - 2),
             # degrees of freedom are always integer
             estimate = ggstatsplot::specify_decimal_p(x = c$estimate[[1]], k),
             LL = ggstatsplot::specify_decimal_p(x = c_ci$conf.low[[1]], k),
@@ -187,11 +194,11 @@ subtitle_ggscatterstats <-
               k,
               p.value = TRUE
             ),
-            n = nrow(x = data)
+            n = sample_size
           )
         )
 
-      #-------------------------------------------------- robust percentage bend --------------------------------------------------------
+      #---------------------- robust percentage bend ----------------------------
     } else if (type == "robust" || type == "r") {
       # running robust correlation
       rob_res <- robcor_ci(
@@ -252,6 +259,76 @@ subtitle_ggscatterstats <-
           sep = ""
         ))
       }
+    } else if (type == "bayes" || type == "bf") {
+
+      # bayes factor results
+      bf_results <- bf_corr_test(
+        data = data,
+        x = x,
+        y = y,
+        bf.prior = bf.prior,
+        caption = NULL,
+        output = "results"
+      )
+
+      # preparing the subtitle
+      subtitle <- base::substitute(
+        expr =
+          paste(
+            "Pearson's ",
+            italic("r"),
+            "(",
+            df,
+            ")",
+            " = ",
+            estimate,
+            # ", 95% CI [",
+            # LL,
+            # ", ",
+            # UL,
+            # "], ",
+            ", log"["e"],
+            "(BF"["10"],
+            ") = ",
+            bf,
+            ", Prior width = ",
+            bf_prior,
+            ", ",
+            # ", log"["e"],
+            # "(error) = ",
+            # bf_error,
+            # "% , ",
+            italic("n"),
+            " = ",
+            n
+          ),
+        env = base::list(
+          # degrees of freedom are always integer
+          df = pearson_r_res$parameter[[1]],
+          estimate = ggstatsplot::specify_decimal_p(x = pearson_r_res$estimate[[1]], k),
+          # LL = ggstatsplot::specify_decimal_p(x = pearson_r_res$conf.int[1][[1]], k),
+          # UL = ggstatsplot::specify_decimal_p(x = pearson_r_res$conf.int[2][[1]], k),
+          bf = ggstatsplot::specify_decimal_p(
+            x = log(
+              x = bf_results$bf[[1]],
+              base = exp(1)
+            ),
+            k = 1
+          ),
+          bf_prior = ggstatsplot::specify_decimal_p(
+            x = bf_results$bf.prior[[1]],
+            k = 3
+          ),
+          # bf_error = ggstatsplot::specify_decimal_p(
+          #   x = log(
+          #     x = bf_results$error[[1]],
+          #     base = exp(1)
+          #   ),
+          #   k = 1
+          # ),
+          n = sample_size
+        )
+      )
     }
 
     # return the subtitle